requirements.txt
alembic==1.4.3
amqp==5.0.1
autocorrect==0.3.0
Babel==2.8.0
backcall==0.2.0
billiard==3.6.3.0
boto3==1.16.5
boto3-stubs==1.16.9.0
botocore==1.19.5
cachalot==1.5.0
celery==5.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
decorator==4.4.2
Django==1.11.1
django-crispy-forms==1.6.1
django-filter==2.0.0
django-guardian==1.5.0
django-js-asset==1.2.2
django-mptt==0.9.0
django-nocaptcha-recaptcha==0.0.19
django-registration==2.2
django-rest-swagger==2.1.2
django-storages==1.5.2
django-widget-tweaks==1.4.1
djangorestframework==3.6.2
docutils==0.16
drf-tracking==1.2.0
ecdsa==0.14.1
et-xmlfile==1.0.1
fasttext==0.9.1
Flask==1.1.2
Flask-Jsonpify==1.5.0
Flask-Migrate==2.5.3
flask-serialize==1.4.0
Flask-SQLAlchemy==2.4.4
flower==0.9.3
gevent==1.3.7
greenlet==0.4.17
gunicorn==20.0.4
idna==2.10
importlib-metadata==2.0.0
ipdb==0.13.4
ipython==7.16.1
ipython-genutils==0.2.0
isoweek==1.3.3
itsdangerous==1.1.0
itypes==1.2.0
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.17.0
jsonify==0.5
jsonpickle==1.4.1
kiwisolver==1.2.0
kombu==5.0.2
langdetect==1.0.7
langid==1.1.6
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.8.0
matplotlib==3.3.1
Morfessor==2.0.1
nltk==3.2.2
numpy==1.19.1
oauthlib==3.1.0
openapi-codec==1.3.2
openpyxl==3.0.4
pandas==1.1.1
parso==0.7.1
permissive-dict==1.0.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
polyglot==16.7.4
prompt-toolkit==3.0.8
psycopg2==2.7.1
psycopg2-binary==2.8.6
ptyprocess==0.6.0
pyasn1==0.4.8
pybind11==2.6.0
pycld2==0.31
pyenchant==1.6.8
pyflakes==1.5.0
Pygments==2.7.2
PyICU==2.3.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
python-jose==3.2.0
python-keycloak==0.22.0
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
regex==2020.7.14
requests==2.24.0
requests-oauthlib==0.7.0
rsa==4.6
s3transfer==0.3.3
scikit-learn==0.23.2
scipy==1.5.3
simplejson==3.17.2
six==1.15.0
snowballstemmer==1.2.1
SQLAlchemy==1.3.20
stop-words==2015.2.23.1
tabulate==0.8.7
threadpoolctl==2.1.0
tinydb==4.2.0
tinydb-smartcache==2.0.0
tornado==5.1.1
tqdm==4.48.2
traitlets==4.3.3
Turkish-Deasciifier @ git+https://github.com/emres/turkish-deasciifier@5d8eddf63bfb8e4199c3c11cc540bc42ed442f4d
unicode-tr==0.6.1
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.10
vine==5.0.0
wcwidth==0.2.5
Werkzeug==1.0.1
xlrd==1.2.0
XlsxWriter==1.1.2
zipp==3.4.0
